Multiple changes:
* Implement caching via md5 hashes to prevent multiple downloads and save
  space.
* Add --subreddit command line option

+++ pr0nit.py
@@ -10,6 +10,7 @@ Author: Tom Dignan <tom@tomdignan.com>
 Contributors: James Clarke 
 """
 
+import md5
 import argparse
 import uuid
 import os
@@ -18,13 +19,15 @@ import urllib2
 import time
 import sys
 
+
 if sys.platform == 'darwin':
     from Foundation import NSAppleScript
     from Cocoa import NSApplication
 
 
+DEBUG = False
 WALLPAPER_CACHE_DIR = "%s/.wallpaper" % os.getenv("HOME")
-DEFAULT_SUBREDDIT = "r/earthporn" # try also, r/usaporn
+DEFAULT_SUBREDDIT = "earthporn" # try also, r/usaporn
 DEFAULT_WALLPAPER_CMD = "feh --bg-scale"
 DEFAULT_FRAME_SPEED = 60
 
@@ -52,14 +55,14 @@ class RedditWallpaperSetter(object):
         """
         try:
             wallpaper_urls = self._get_wallpaper_urls()
-        except urllib2.HTTPError, e:
+        except urllib2.HTTPError as e:
             print "Retrying in 10 seconds: %r" % e
             time.sleep(10)
             self.run()
             return
 
         for url in wallpaper_urls:
-            path = self._download_wallpaper(url)            
+            path = self._cache_wallpaper(url)            
             self._set_wallpaper(path)
             time.sleep(self.frame_speed)
 
@@ -85,16 +88,26 @@ class RedditWallpaperSetter(object):
         return wallpaper_urls
 
 
-    def _download_wallpaper(self, url):
+    def _cache_wallpaper(self, url):
         """ 
-        Saves the wallpaper at the target `url` in `cache_dir`
+        Saves the wallpaper at the target `url` in `cache_dir` but only if
+        there is no file in the cache which appears to be the target wallpaper
+        already. The heuristic used is an md5sum of the target wallpaper's url.
         """ 
         file_in = urllib2.urlopen(url)
-        path = self.cache_dir + os.path.sep + str(uuid.uuid4()) + ".jpg"
-        file_out = open(path, "w")
-        file_out.write(file_in.read())
-        file_out.close()
-        file_in.close()
+        summer = md5.new()
+        summer.update(url)
+        filename = summer.hexdigest() + ".jpg"
+        path = os.path.join(self.cache_dir,  filename)
+
+        # only create the file if it does not already exist in the cache.
+        if not os.path.isfile(path): 
+            file_out = open(path, "w")
+            file_out.write(file_in.read())
+            file_out.close()
+            file_in.close()
+        else:
+            print "cache hit: %r %r" % (url, path)
 
         return path
 
@@ -110,7 +123,6 @@ class RedditWallpaperSetterLinux(RedditWallpaperSetter):
 
 
 class RedditWallpaperSetterXFCE4(RedditWallpaperSetter):
-    
     def _set_wallpaper(self, path):
         for i in range(self.monitors):
             os.system("xfconf-query -c xfce4-desktop -p "  
@@ -131,6 +143,9 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="""Set the wallpaper from
  reddit""")
 
+    parser.add_argument("--subreddit", metavar="<name>", type=str, nargs=1,
+        default=DEFAULT_SUBREDDIT, help="Target subreddit to scrape for "
+        " wallpaper.")
 
     parser.add_argument("--monitors", metavar="<number>", type=int, nargs=1,
         default=1, help="Number of monitors in your setup (XFCE4 only)")
@@ -145,28 +160,26 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
     
-    get_arg = lambda x : x[0] if not x is None and not isinstance(x, int) else x
+    get_arg = lambda x : x[0] if not x is None and not isinstance(x, int) \
+                        and not isinstance(x, basestring) else x
+
+
     monitors = get_arg(args.monitors)
     platform = get_arg(args.platform)
     frame_speed = get_arg(args.frame_speed)
+    subreddit = "r/" + get_arg(args.subreddit)
 
 
     if platform == "darwin":
         # Set the activation policy to NSApplicationActivationPolicyAccessory
         # so we don't show the Python dock icon when using PyObjC.
         NSApplication.sharedApplication().setActivationPolicy_(2)
-        wallpaper_setter = RedditWallpaperSetterOSX(DEFAULT_SUBREDDIT,
-                                                    WALLPAPER_CACHE_DIR,
-                                                    frame_speed,
-                                                    monitors=monitors)
+        wallpaper_setter = RedditWallpaperSetterOSX(subreddit, WALLPAPER_CACHE_DIR,
+                                                    frame_speed, monitors=monitors)
     elif platform == "xfce4":
-        wallpaper_setter = RedditWallpaperSetterXFCE4(DEFAULT_SUBREDDIT,
-                                                      WALLPAPER_CACHE_DIR,
-                                                      frame_speed,
-                                                      monitors=monitors)
+        wallpaper_setter = RedditWallpaperSetterXFCE4(subreddit, WALLPAPER_CACHE_DIR,
+                                                      frame_speed, monitors=monitors)
     else:
-        wallpaper_setter = RedditWallpaperSetterLinux(DEFAULT_SUBREDDIT,
-                                                      WALLPAPER_CACHE_DIR,
-                                                      frame_speed,
-                                                      monitors=monitors)
+        wallpaper_setter = RedditWallpaperSetterLinux(subreddit, WALLPAPER_CACHE_DIR,
+                                                      frame_speed, monitors=monitors)
     wallpaper_setter.run()
